refactor: move soundengineaccessor to accessor package
